refactor: moving PlaceInfoContext from 'src\components\EditPage\temp_data\trip_reducer.jsx' to 'src\contexts\PlaceInfoContext.jsx'
