+++ src/contexts/PlaceInfoContext.jsx
@@ -1,29 +1,34 @@
-//////////////////// reducer ////////////////////
-
-import moment from 'moment';
 import { useContext, createContext, useReducer } from 'react';
 
-const PlaceInfoContext = createContext();
-const PlaceInfoDispatchContext = createContext();
+//////////////////// actions ////////////////////
 
 export const placeInfo_actions = {
-  // GET_PLACE_INFO: 'GET_PLACE_INFO', // 取得搜尋景點的資訊
-  SET_PLACE_INFO: 'SET_PLACE_INFO', // 暫存搜尋景點的資訊
-  DELETE_PLACE_INFO: 'DELETE_PLACE_INFO', // 刪除搜尋景點的資訊
+  SET_PLACE_INFO: 'SET_PLACE_INFO', // 暫存景點搜尋結果 (from Google Map Autocomplete)
+  DELETE_PLACE_INFO: 'DELETE_PLACE_INFO', // 刪除景點搜尋結果
 };
+
+//////////////////// reducer ////////////////////
 
 function placeInfoReducer(placeInfo, action) {
   switch (action.type) {
-    case placeInfo_actions.SET_PLACE_INFO:
-      const data = action.payload;
-      return { ...placeInfo, ...data };
-    case placeInfo_actions.DELETE_PLACE_INFO:
+    case placeInfo_actions.SET_PLACE_INFO: {
+      const placeData = action.payload;
+      return { ...placeData };
+    }
+    case placeInfo_actions.DELETE_PLACE_INFO: {
       return {};
-    default:
+    }
+    default: {
       console.log('placeInfo dispatch error');
       break;
+    }
   }
 }
+
+//////////////////// context ////////////////////
+
+const PlaceInfoContext = createContext();
+const PlaceInfoDispatchContext = createContext();
 
 export function PlaceInfoProvider({ children }) {
   const [placeInfo, placeInfoDispatch] = useReducer(placeInfoReducer, {});
